Successfully implemented the back button. Given new name to the app SeePain.
