+++ pages/menu_page/seepain_info/kv/drawing.kv
@@ -1,0 +1,4 @@
+<PaintWidget>:
+    size_hint: 1, 0.9
+    pos_hint: {"top": 1}
+
