--- pain2d_info/kv/drawing.kv
@@ -1,4 +1,0 @@
-<PaintWidget>:
-    size_hint: 1, 0.9
-    pos_hint: {"top": 1}
-
